+++ packages/common-ethereum/CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Missing imports for multi abi `codegen` (#169)
-
 ### Added
 - Parent field to manifest for project upgrades (#148)
 - Light Log types to codegen (#170)
+### Changed
+- Update model with `types-core`
+
 
 ## [2.3.0] - 2023-09-12
 ### Changed